+++ test/internal.js
@@ -1,6 +1,6 @@
 var test = require('tape')
-var Mingo = require('../dist/mingo')
-var _ = Mingo._internal()
+var mingo = require('../dist/mingo')
+var _ = mingo._internal()
 
 // TODO: add some more tests
 test('Test isEqual', function (t) {
@@ -13,14 +13,14 @@ test('Test isEqual', function (t) {
     [NaN, 1, false],
     [[1, 2], [1, 2], true],
     [[2, 1], [1, 2], false],
-    [[1, 'a', {a: /b/}], [1, 'a', {a: RegExp('b')}], true],
+    [[1, 'a', {a: /b/}], [1, 'a', {a: new RegExp('b')}], true],
     [null, undefined, false],
     [new Date(2003, 10, 1), new Date(2003, 10, 1), true],
     [{date: {year: 2013, month: 9, day: 25}}, {date: {year: 2013, month: 9, day: 25}}, true],
     [function () {}, function () {}, false],
     [Object.prototype.toString, Object.prototype.toString, true]
   ]
-  sample.forEach(function (arr, i) {
+  sample.forEach(function (arr) {
     t.equal(_.isEqual(arr[0], arr[1]), arr[2])
   })
   t.end()